Исправлен тип данных, используемых при вводе суммы в консоле

+++ src/FinalTask3.java
@@ -8,7 +8,7 @@ public class FinalTask3 {
         System.out.print("Введит текущий курс доллара: ");
         double rate = scan.nextDouble();
         System.out.print("Введите сумму в рублях: ");
-        double amount = scan.nextInt();
+        double amount = scan.nextDouble();
         BigDecimal total = new BigDecimal(amount/rate).setScale(2,RoundingMode.HALF_UP);
         System.out.printf("Курс доллара = %.2f, Количество рублей : %.0f. Итого: %.2f долларов", rate, amount, total);
         scan.close();
